Hydrate routines data on plan page
- colocation을 신경 써보았다.
