+++ types/routine.ts
@@ -1,4 +1,3 @@
-import { RoutineCheck } from 'src/constants/routine';
 import { Category } from 'types/category';
 
 export type Routine = {
@@ -6,19 +5,4 @@ export type Routine = {
   name: string;
   repeatDays: number[];
   category: Category;
-};
-
-export type DailyRoutine = {
-  id: string; // uuid
-  name: string;
-  routineCheck: RoutineCheck | null; // null은 0으로 취급한다.
-  category: Category;
-};
-
-export type DailyRoutineTab = {
-  id: RoutineCheck;
-  name: string;
-  routines?: DailyRoutine[]; // server state
-  checkable?: boolean;
-  updatable?: boolean;
 };
